+++ modules/Accounting/includes/Dashboard.inc.php
@@ -16,8 +16,6 @@
  */
 function DashboardDefaultAccounting()
 {
-	require_once 'ProgramFunctions/DashboardModule.fnc.php';
-
 	$profile = User( 'PROFILE' );
 
 	$data = '';
@@ -27,7 +25,7 @@ function DashboardDefaultAccounting()
 		$data = DashboardAccountingAdmin();
 	}
 
-	return DashboardModule( 'Accounting', $data );
+	return $data;
 }
 
 if ( ! function_exists( 'DashboardAccountingAdmin' ) )