+++ modules/Scheduling/includes/Dashboard.inc.php
@@ -16,8 +16,6 @@
  */
 function DashboardDefaultScheduling()
 {
-	require_once 'ProgramFunctions/DashboardModule.fnc.php';
-
 	$profile = User( 'PROFILE' );
 
 	$data = '';
@@ -27,7 +25,7 @@ function DashboardDefaultScheduling()
 		$data = DashboardSchedulingAdmin();
 	}
 
-	return DashboardModule( 'Scheduling', $data );
+	return $data;
 }
 
 if ( ! function_exists( 'DashboardSchedulingAdmin' ) )